fix and shrink
